Support scenario outlines
Previously, a linear line-by-line processing of scenario specifications
was assumed, which was helpful, but for supporting scenario outlines,
we need the ability for running the steps again with new data. For
this reason the processing and its buffering was refactored for
supporting rerunning steps at any time.

It was also assumed tables can be parsed simply via regular
expressions, which worked fine for simple cases, but in the case of
scenario outlines we need to reference values from a table by a named
column, and we needed a full support for getting data a row at a time.

A large part of scenario outlines works via a simple reusable
search-and-replace algorithm that expand tokens with data. Reusable
helpers were created for this purpose: expand_token() and
expand_tokens() and these are called from call_step() if we're running
steps on a scenario outline.

Example tests for scenario outlines were added to examples/tests/,
please see the "bank.feature".