Use macros for testing success within steps
Previously the return value from match_expect was relied upon for
knowing whether a step was successful or not, but this didn't give
an easy way of printing file:line of a failure within a step. With
macros we get this, and this will be useful for setting breakpoints
in a debugger.