Use pipes for passing pass/fail to main process
We use pipes instead of the exit status because the child processes
may exit any number of ways, while with successful pipe reads, we can
be quite certain a pass is a pass and a fail is a fail, and any other
situation is a fail as well. We can also later use pipes for passing
detailed status information.