Fix GNU make compatibility
Originally written for BSD make, made some fixes for GNU make.

The examples/Makefile.in is an example for building a project with
optional tests. It could use for-loops and various fancy things,
but for avoiding a need for using makefile generators, and for
avoiding having a dependency on GNU make, only the simplest idioms
are used.